[update] generic improvements

+++ terminal.py
@@ -1,13 +1,25 @@
 #!/usr/bin/python
 
 from typing import Optional, Any, Dict, List, Union
-from subprocess import run, Popen, PIPE, CalledProcessError
+from subprocess import run, Popen, PIPE, STDOUT, CalledProcessError
 from shlex import split as parse_params
 from re import sub
 from getpass import getpass
 import os
 import pwd
 import glob
+import copy
+from utilities import (
+	exec_exit, 
+	is_empty,
+	array_clear,
+	sprintf,
+	printinf,
+	printalr,
+	printerr,
+	printsuc,
+	import_module_error
+)
 
 try:
     from prompt_toolkit import PromptSession
@@ -22,17 +34,6 @@ try:
 	from pynput import keyboard
 except ImportError:
 	import_module_error("pynput")
-
-from utilities import (
-	exec_exit, 
-	is_empty,
-	array_clear,
-	sprintf,
-	printinf,
-	printalr,
-	printerr,
-	printsuc
-)
 from echo import Echo
 from thread_maid import ThreadMaid
 
@@ -40,20 +41,24 @@ BASE = os.path.dirname(__file__)
 PID = str(os.getpid()).strip()
 LOCK_FILE = "/tmp/powerconsole_hotkey.lock"
 
+# commands_thread = ThreadMaid()
 cuid = os.getuid()
 cugid = os.getgid()
 is_super_user = False if cuid != 0 else True
 e = os.environ.copy()
-aliases = []
 git_commands = {
-	"add": WordCompleter(["."]),
-	"commit": WordCompleter(["-m"]),
-	"checkout": WordCompleter([
-		"-b"
-		"master",
-		"main",
-		"development"
-	]),
+	"add": {
+		".": None
+	},
+	"commit": {
+		"-m": None
+	},
+	"checkout": {
+		"-b": None,
+		"master": None,
+		"main": None,
+		"development": None
+	},
 	"status": None,
 	"branch": None,
 	"log": None
@@ -65,50 +70,64 @@ completer = {
 	"reload-conf": None,
 	"reload-config": None,
 	"reload-env": None,
-	"ln": WordCompleter(["-s"])
+	"ln": {
+		"-s": None
+	},
+	"cp": {
+		"-r": None
+	},
+	"rm": {
+		"-r": None
+	},
+	"cd": {},
+	"mv": {}, 
+	"nano": {}, 
+	"vi": {}, 
+	"vim": {}, 
+	"subl": {}
 }
-commands_thread = ThreadMaid()
-
-
-def escalate():
-	try:
-		password = getpass(f"[sudo] password for {user}: ")
-		run(
-			["sudo", "-i"], 
-			input=password, 
-			text=True, 
-			check=True, 
-			capture_output=True
-		)
-		is_super_user = True
-		# os.setuid(0)
-		print("You are now super user")
-	except PermissionError as pe:
-		print(f"Unable to change uid: {pe}")
-	except Exception as err:
-		print(f"Unable to escalate privileges: {err}")
-
-
-def return_to_user():
-	try:
-		os.seteuid(cuid)
-		is_super_user = False
-	except Exception:
-		print("Unable to return to original user")
 
 
 def get_username():
     return pwd.getpwuid(cuid).pw_name
 
 
-user = get_username()
+def escalate():
+	global is_super_user
+
+	try:
+		password = getpass(f"[sudo] password for {get_username()}: ")
+		run(
+            ['sudo', '-S', '-v'], 
+            input=password, 
+            text=True, 
+            check=True, 
+            capture_output=True
+        )
+		is_super_user = True
+		printinf("You are now super user")
+	except PermissionError as pe:
+		printerr(f"Unable to change uid: {pe}")
+	except Exception as err:
+		printerr(f"Unable to gain root privileges: {err}")
+
+
+def deescalate():
+	global is_super_user
+
+	try:
+		run(['sudo', '-k'])
+		is_super_user = False
+		printinf(f"You are now back to user {get_username()}")
+	except Exception:
+		printerr("Unable to return to original user")
 
 
 def get_commands():
 	process = Popen(
         ["ls", "/usr/bin"],
-        stdin=PIPE, 
-        stderr=PIPE, 
+        stdin=PIPE,
+        stderr=PIPE,
         stdout=PIPE
     )
 	output, error = process.communicate()
@@ -117,8 +136,8 @@ def get_commands():
 
 def get_local_commands():
 	process = Popen(
-        ["ls", f"/home/{user}/.local/bin"],
-        stdin=PIPE, 
+        ["ls", f"/home/{get_username()}/.local/bin"],
+        stdin=PIPE,
         stderr=PIPE, 
         stdout=PIPE
     )
@@ -167,49 +186,55 @@ def get_git_repo() -> str:
 
 
 def autocomplete_cwd():
-	git_commands["push"] = WordCompleter([
-		"origin", 
-		"master",
-		"main",
-		"development",
-		get_git_repo()
-	]),
-	git_commands["pull"] = WordCompleter([
-		"origin",
-		"master",
-		"main",
-		"development",
-		get_git_repo()
-	]),
-	completer["git"] = NestedCompleter(git_commands)
+	global completer, git_commands
 
-	completer["cd"] = WordCompleter(get_files())
-	completer["cp"] = WordCompleter(["-r", get_files()])
-	completer["mv"] = WordCompleter(get_files())
-	completer["rm"] = WordCompleter(["-r", "-f", get_files()])
-	completer["nano"] = WordCompleter(get_files())
-	completer["vi"] = WordCompleter(get_files())
-	completer["vim"] = WordCompleter(get_files())
-	completer["subl"] = WordCompleter(get_files())
-	completer["code"] = WordCompleter(get_files())
+	suggestions = copy.deepcopy(completer)
+
+	repo = get_git_repo()
+
+	git_commands["push"] = {
+		"origin": {}, 
+		"master": None,
+		"main": None,
+		"development": None
+	}
+	git_commands["pull"] = {
+		"origin": {
+			"master": None,
+			"main": None,
+			"development": None
+		}, 
+		"master": None,
+		"main": None,
+		"development": None
+	}
+	git_commands["push"][repo] = None
+	git_commands["pull"][repo] = None
+	git_commands["push"]["origin"][repo] = None
+	git_commands["pull"]["origin"][repo] = None
+
+	suggestions["git"] = git_commands
 
 	# Add /usr/bin/ commands
-	"""
 	for c in get_commands():
-		completer[c] = None
-	"""
+		suggestions[c] = None	
 
-	# Add local commands
-	"""
+	# Add local commands	
 	for c in get_local_commands():
-		completer[c] = None
-	"""
+		suggestions[c] = None	
 
 	# Add files of the current directory
 	for f in get_files():
-		completer[f] = None
+		suggestions[f] = None
 
-	return NestedCompleter(completer)
+		for c in ["cd", "cp", "rm", "mv", "nano", "vi", "vim", "subl"]:
+			if c in suggestions:
+				if suggestions[c] is None:
+					suggestions[c] = {}
+				
+				suggestions[cmd][f] = None
+
+	return NestedCompleter.from_nested_dict(suggestions)
 
 
 def prompt(ppt):
@@ -225,7 +250,8 @@ def prompt(ppt):
 	        return session.prompt(
 	        	ppt,
 	        	auto_suggest=AutoSuggestFromHistory(),
-	        	complete_while_typing=True
+	        	complete_while_typing=True,
+	        	wrap_lines=True
 	        )
 	    except KeyboardInterrupt:
 	    	return ""
@@ -235,7 +261,7 @@ def prompt(ppt):
 	    	print(f"Unknown exception in prompt: {err}")
 
 
-bash_history = f"/home/{user}/.bash_history"
+bash_history = f"/home/{get_username()}/.bash_history"
 history = FileHistory(bash_history)
 
 
@@ -253,19 +279,20 @@ keyboard_listener = None
 def open_new_tab():
 	run(
 		[
-			"gnome-terminal", 
-			"--tab", 
-			f"--working-directory={os.getcwd()}", 
-			"--", 
-			"python", 
+			"gnome-terminal",
+			"--tab",
+			f"--working-directory={os.getcwd()}",
+			"--",
+			"python",
 			f"{BASE}/terminal.py"
-		], 
+		],
 		env=os.environ.copy()
 	)
 
 	return True
 
 
+"""
 def handle_commands():
 	def for_canonical(f):
 		return lambda k: f(l.canonical(k))
@@ -286,36 +313,54 @@ commands_thread.setup(target=handle_commands)
 def create_commands_thread():
 	commands_thread = ThreadMaid()
 	commands_thread.setup(target=handle_commands)
+"""
+
+
+def kill_lock():
+	if os.path.exists(LOCK_FILE):
+		content = ""
+
+		with open(LOCK_FILE, "r") as f:
+			content = f.read().strip()
+        
+		# Are we already the master?
+		if PID == content:
+			return True
+
+		# Is the master still alive?
+
+		# This logic is broken
+		try:
+			"""
+			printinf("The master tab got killed, so this tab gained its powers")
+	
+			# needs to check if the process is still alive
+			# before trying to kill it
+			os.kill(int(content), 0)
+			"""
+			return False
+		except (OSError, ValueError):
+			try:
+				os.remove(LOCK_FILE)
+			except:
+				pass
 
 
 def is_master_tab():
-    if os.path.exists(LOCK_FILE):
-        with open(LOCK_FILE, "r") as f:
-            content = f.read().strip()
-        
-        # Are we already the master?
-        if PID == content:
-            return True
-        
-        # Is the master still alive?
-        try:
-            os.kill(int(content), 0)
-            return False 
-        except (OSError, ValueError):
-            try:
-                os.remove(LOCK_FILE)
-            except:
-                pass
+	kill_lock()
 
-    try:
-        with open(LOCK_FILE, "x") as f:
-            f.write(str(os.getpid()))
-        
-        create_commands_thread()
-        commands_thread.run()
-        return True
-    except (FileExistsError, OSError):
-        return False
+	try:
+		with open(LOCK_FILE, "x") as f:
+			f.write(str(os.getpid()))
+
+		"""
+		create_commands_thread()
+		commands_thread.run()
+		"""
+		return True
+	except (FileExistsError, OSError):
+		return False
+
 
 is_master = is_master_tab()
 
@@ -355,9 +400,9 @@ while True:
 			]),
 			# Replacements
 			user_marker, # 0
-			user, # 1
-			dir_, # 2
-			" 🖝  " # 3
+			get_username(), # 1
+			dir_, # 2,
+			" 🖝  "# 3
 		)
 	)
 
@@ -386,14 +431,18 @@ while True:
 
 	try:
 		if cmd == "exit":
-			if not is_empty(keyboard_listener):
-				keyboard_listener.stop()
+			if is_super_user:
+				deescalate()
+				continue
+
+			# if not is_empty(keyboard_listener):
+			# 	keyboard_listener.stop()
 
 			if os.path.exists(LOCK_FILE) and is_master:
 				os.remove(LOCK_FILE)
 
-			commands_thread.halt()
-			exec_exit()
+			# commands_thread.halt()
+			exec_exit(True)
 
 		elif cmd == "pid":
 			print(PID)
@@ -403,10 +452,14 @@ while True:
 			if args[0] == "tab":
 				if args[1] == "master":
 					print("This tab is the master") if is_master else print("This tab is a clone")
-
 			continue
 
-		elif cmd in ["reload-config", "reload-conf", "reload-env"]:
+		elif cmd == "new" and argsvalid:
+			if args[0] == "tab":
+				open_new_tab()
+			continue
+
+		elif cmd in ["reload-config", "reload-conf", "reload-env", "reload"]:
 			e = os.environ.copy()
 			print("Done")
 			continue
@@ -432,6 +485,10 @@ while True:
 				escalate()
 				continue
 
+		elif cmd == "whoami":
+			if is_super_user:
+				printalr("You are executing commands as root, be careful!")
+
 		# Executing bash scripts
 		elif cmd.startswith("./"):
 			script = cmd.replace("./", "")
@@ -443,9 +500,8 @@ while True:
 			cmd = "bash"
 			args.append(os.getcwd() + "/" + script)
 
-		# print([cmd, "-E", *args])
 		if is_super_user:
-			run([cmd, "-E", *args], env=e)
+			run(["sudo", "-E", cmd, *args], env=e)
 		else:
 			run([cmd, *args], env=e)
 	except CalledProcessError as cpe:
@@ -453,9 +509,13 @@ while True:
 		continue
 	except Exception as err:
 		print(f"Error occurred: {err}")
+
+		if os.path.exists(LOCK_FILE) and is_master:
+			os.remove(LOCK_FILE)
+
 		continue
 	except KeyboardInterrupt:
 		if is_super_user:
-			return_to_user()
+			deescalate()
 
 		continue
